Add download model API in compressor v2 client
